refactor: 페이지 응답 값 Generic type 적용 (#46)
