+++ winwin-be-api/src/main/java/com/dpm/winwin/api/common/response/dto/GlobalPageResponseDto.java
@@ -1,22 +1,21 @@
-package com.dpm.winwin.api.post.dto.response;
+package com.dpm.winwin.api.common.response.dto;
 
 import java.util.List;
 import org.springframework.data.domain.Page;
 
-public record PostCustomizedListResponse(
-    List<PostCustomizedResponse> content,
+public record GlobalPageResponseDto<T>(
+    List<T> content,
     long totalElements,
     int totalPages,
     boolean hasNextPages
 ) {
 
-    public static PostCustomizedListResponse of(Page page) {
-        return new PostCustomizedListResponse(
+    public static <T> GlobalPageResponseDto<T> of(Page<T> page) {
+        return new GlobalPageResponseDto<T>(
             page.getContent(),
             page.getTotalElements(),
             page.getTotalPages(),
             page.hasNext()
         );
     }
-
 }
