+++ winwin-be-api/src/main/java/com/dpm/winwin/api/post/controller/PostController.java
@@ -4,7 +4,8 @@ import com.dpm.winwin.api.common.response.dto.BaseResponseDto;
 import com.dpm.winwin.api.post.dto.request.PostAddRequest;
 import com.dpm.winwin.api.post.dto.request.PostUpdateRequest;
 import com.dpm.winwin.api.post.dto.response.PostAddResponse;
-import com.dpm.winwin.api.post.dto.response.PostCustomizedListResponse;
+import com.dpm.winwin.api.common.response.dto.GlobalPageResponseDto;
+import com.dpm.winwin.api.post.dto.response.PostCustomizedResponse;
 import com.dpm.winwin.api.post.dto.response.PostListResponse;
 import com.dpm.winwin.api.post.dto.response.PostMethodsResponse;
 import com.dpm.winwin.api.post.dto.response.PostReadResponse;
@@ -13,7 +14,6 @@ import com.dpm.winwin.api.post.service.PostService;
 import com.dpm.winwin.domain.repository.post.dto.request.PostCustomizedConditionRequest;
 import com.dpm.winwin.domain.repository.post.dto.request.PostListConditionRequest;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -33,18 +33,18 @@ public class PostController {
   private final PostService postService;
 
   @GetMapping
-  public BaseResponseDto<Page<PostListResponse>> getPostList(
+  public BaseResponseDto<GlobalPageResponseDto<PostListResponse>> getPostList(
       PostListConditionRequest condition, Pageable pageable
   ) {
-    Page<PostListResponse> response = postService.getPosts(condition, pageable);
+    GlobalPageResponseDto<PostListResponse> response = postService.getPosts(condition, pageable);
     return BaseResponseDto.ok(response);
   }
 
   @GetMapping("/custom")
-  public BaseResponseDto<PostCustomizedListResponse> getCustomPostList(
+  public BaseResponseDto<GlobalPageResponseDto<PostCustomizedResponse>> getCustomPostList(
       PostCustomizedConditionRequest condition, Pageable pageable) {
     Long tempMemberId = 1L;
-    PostCustomizedListResponse response = postService
+    GlobalPageResponseDto<PostCustomizedResponse> response = postService
         .getPostsCustomized(tempMemberId, condition, pageable);
     return BaseResponseDto.ok(response);
   }
